requirements.txt
# AI Studio — additional requirements (on top of voice-pro deps)
# Install: pip install -r requirements.txt

# Video generation
diffusers>=0.29.0
accelerate>=0.26.0
transformers>=4.40.0
imageio[ffmpeg]>=2.34.0
imageio-ffmpeg>=0.4.9

# Audio generation (Meta AudioCraft)
# Note: audiocraft requires Python 3.9+ and PyTorch 2.x
audiocraft>=1.3.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
rich>=13.0
huggingface-hub>=0.27.0

# Language detection (for auto-translation of prompts)
lingua-language-detector>=2.0.0
deep-translator>=1.11.0
